+++ backend/substrapp/models/metric.py
@@ -15,7 +15,6 @@ class Metric(TimeStamped):
     """Storage Metric table"""
 
     key = models.UUIDField(primary_key=True, default=uuid.uuid4, editable=False)
-    validated = models.BooleanField(default=False, blank=True)
     description = models.FileField(
         storage=settings.METRICS_STORAGE, max_length=500, upload_to=upload_to, blank=True, null=True
     )  # path max length to 500 instead of default 100
@@ -31,4 +30,4 @@ class Metric(TimeStamped):
         super().save(*args, **kwargs)
 
     def __str__(self) -> str:
-        return f"Metric with key {self.key} with validated {self.validated}"
+        return f"Metric with key {self.key}"