+++ backend/substrapp/views/metric.py
@@ -107,7 +107,7 @@ class MetricViewSet(mixins.CreateModelMixin, PaginationMixin, GenericViewSet):
         description_file = tempfile.TemporaryFile()
         description_file.write(content)
 
-        instance, created = Metric.objects.update_or_create(key=key, validated=True)
+        instance, created = Metric.objects.update_or_create(key=key)
         instance.description.save("description.md", description_file)
 
         return instance